move out rails, satin processor, and stitches
